pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contextualize-docs"
version = "0.1.0"
description = "Agent-native docs compiler for Contextualize"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0,<3.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "google-genai>=1.0.0",  # optional: only needed if using direct Gemini SDK
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.scripts]
contextualize-docs = "contextualize_docs.cli:main"

[tool.pytest.ini_options]
testpaths = ["contextualize_docs/tests"]
asyncio_mode = "auto"
